WIP: Move typed settings dictionary to CoreConfiguration
This lets us put enums from CoreConfiguration into the dictionary, to enable their usage from the command line
